tests passed, with one example for docstring format
